Progress on refactoring simulation.c
I have made significant changes to the body of the loop in simulation_step.
Previously, there were several if statements, extending over a hundred
lines or so, which I have replaced with a switch-case over the entity type.
Each case is quite short, calling util functions to implement the
operation, such as human birth/death or zombies infecting humans.

The util functions still need to be implemented.